blacklist word added

+++ src/commands.js
@@ -32,10 +32,13 @@ module.exports = async function (msg) {
   */
 
   //delete messages
-  /*if (msg.content.includes("YAY")) {
-    let input = msg.content.split(" ").slice(1).join(" "); // Removes the prefix
-    msg.delete(); // Deletes the message
-    //msg.channel.send(input); //.then(msg=>msg.delete({timeout:"5000"}) <- if you want delete it with delay and sends the finished text
+  /*var blackList = ["YAY", "yay", "yAy", "YaY", "yAY", "Yay", "YAy"];
+  for (var i = 0; i < blackList.length; i++) {
+    if (msg.content.includes(blackList[i])) {
+      let input = msg.content.split(" ").slice(1).join(" "); // Removes the prefix
+      msg.delete(); // Deletes the message
+      //msg.channel.send("Stop Saying: " + input); //.then(msg=>msg.delete({timeout:"5000"}) <- if you want delete it with delay and sends the finished text
+    }
   }*/
 
   //stocks
